test dev

+++ database/migrations/2019_11_08_140000_create_evenement_table.php
@@ -23,7 +23,7 @@ class CreateEvenementTable extends Migration
             $table->string('description_evenement');
             $table->date('date_evenement');
             $table->boolean('reccurence');
-            $table->integer('prix');
+            $table->bigInteger('prix');
             $table->date('date_creation');
             $table->integer('user_id');
             $table->timestamps();
